refactor: 유틸함수 분리 후 기존 파일 삭제
